pyproject.toml
[project]
name = "mng-daily-digest"
version = "1.0.0"
description = "Mississippi News Group Daily Digest Newsletter Generator"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}
authors = [
    {name = "Mississippi News Group"}
]
keywords = ["newsletter", "rss", "email", "digest", "mississippi", "news"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: End Users/Desktop",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Communications :: Email",
    "Topic :: Internet :: WWW/HTTP :: Dynamic Content :: News/Diary",
]

dependencies = [
    "feedparser>=6.0.10",
    "requests>=2.31.0",
    "httpx>=0.25.0",
    "python-dateutil>=2.8.2",
    "beautifulsoup4>=4.12.2",
    "lxml>=4.9.3",
    "jinja2>=3.1.2",
    "pyyaml>=6.0.1",
    "python-dotenv>=1.0.0",
    "aiosmtplib>=3.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "responses>=0.23.0",
    "mypy>=1.5.0",
    "types-requests>=2.31.0",
    "types-python-dateutil>=2.8.19",
    "types-beautifulsoup4>=4.12.0",
    "types-PyYAML>=6.0.12",
]

[project.scripts]
mng-digest = "main:main"

[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["providers*", "storage*", "renderer*", "email_sender*", "utils*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]
addopts = "-v --tb=short"

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true
